requirements.txt
# 港股十倍股量化選股系統依賴
streamlit>=1.28.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
requests>=2.31.0
